v2/requirements.txt
pyspark==3.4.1
psycopg2-binary==2.9.9
pandas==2.1.1
cryptography==41.0.4
python-dotenv==1.0.0
click==8.1.7
prometheus-client==0.17.1
structlog==23.1.0
retry==0.9.2
jsonschema==4.19.1
PyYAML==6.0.1
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
isort==5.12.0
mypy==1.5.1
